requirements.txt
Django==6.0.3
django-bootstrap5==24.3
Pillow==10.4.0